Automatically create snippets folder if it doesn't exist.

+++ includes/ConsoleSnippets.php
@@ -2,6 +2,10 @@
 
 $snippetsPath = $this->wire('config')->paths->site.\TracyDebugger::getDataValue('snippetsPath').'/TracyDebugger/snippets/';
 
+if(!is_dir($snippetsPath)) if(!wireMkdir($snippetsPath, true)) {
+    throw new WireException("Unable to create snippets path: $snippetsPath");
+}
+
 if(isset($_POST['deletesnippet']) && file_exists($snippetsPath.$_POST['snippetname'])) {
     unlink($snippetsPath.$_POST['snippetname']);
 }
